Can download demos for accounts with Game Authentication Code [1]
Needed to download demos with Game Authentication Code:
- an <account_name>.lastmatch file with a valid game code for the account
- a Game Authentication Code for the account
- Steam API key
- any account with username + password without Steam Authenticator. This account will be used to download the demo files.

Login on the first account with password will trigger downloading the demos for each account with Game Authentication Code. Afterwards, will download the demos for all the accounts with password through the old method.
Update dependencies. Save timestamp in .json because the .info scoretable file doesn't work with the new protos.


[1] https://developer.valvesoftware.com/wiki/Counter-Strike:_Global_Offensive_Access_Match_History